student_management_system/students/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tortoise-orm>=0.17.0
aiomysql>=0.0.21
pydantic>=1.8.0
email-validator>=1.1.3
python-multipart>=0.0.5
aerich>=0.5.3
python-dotenv>=0.19.0

cryptography>=3.4.7

